# docs/swagger.yaml
basePath: /
definitions:
  handlers.AcknowledgeMessageRequest:
    description: Request to acknowledge (delete) messages
    properties:
      messageIds:
        items:
          type: string
        type: array
    type: object
  handlers.DeliveryBlockedError:
    description: Error response when delivery is blocked for some recipients
    properties:
      blockedRecipients:
        items:
          type: string
        type: array
      code:
        example: ERR_DELIVERY_BLOCKED
        type: string
      description:
        example: 'Blocked recipients: 03abc...'
        type: string
      status:
        example: error
        type: string
    type: object
  handlers.DeviceOut:
    description: Device registration object
    properties:
      active:
        example: true
        type: boolean
      createdAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
      deviceId:
        example: device-abc
        type: string
      fcmToken:
        example: '...abc123'
        type: string
      id:
        example: 1
        type: integer
      lastUsed:
        example: "2024-01-01T12:00:00.000Z"
        type: string
      platform:
        example: ios
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
    type: object
  handlers.ErrorResponse:
    description: Error response with status, code, and description
    properties:
      code:
        example: ERR_INVALID_REQUEST
        type: string
      description:
        example: Invalid request parameters
        type: string
      status:
        example: error
        type: string
    type: object
  handlers.GetPermissionResponse:
    description: Response containing permission details
    properties:
      description:
        example: Permission setting found.
        type: string
      permission:
        $ref: '#/definitions/handlers.PermissionDetail'
      status:
        example: success
        type: string
    type: object
  handlers.ListDevicesResponse:
    description: Response containing list of devices
    properties:
      devices:
        items:
          $ref: '#/definitions/handlers.DeviceOut'
        type: array
      status:
        example: success
        type: string
    type: object
  handlers.ListMessagesRequest:
    description: Request to list messages from a message box
    properties:
      messageBox:
        type: string
    type: object
  handlers.ListMessagesResponse:
    description: Response containing list of messages
    properties:
      messages:
        items:
          $ref: '#/definitions/handlers.MessageOut'
        type: array
      status:
        example: success
        type: string
    type: object
  handlers.ListPermissionsResponse:
    description: Response containing list of permissions
    properties:
      permissions:
        items:
          $ref: '#/definitions/handlers.PermissionDetail'
        type: array
      status:
        example: success
        type: string
      totalCount:
        example: 10
        type: integer
    type: object
  handlers.MessageOut:
    description: Message object returned by listMessages
    properties:
      body:
        example: '{"text":"Hello"}'
        type: string
      createdAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
      messageId:
        example: abc123
        type: string
      sender:
        example: 03abc...
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
    type: object
  handlers.PermissionDetail:
    description: Permission details
    properties:
      createdAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
      messageBox:
        example: inbox
        type: string
      recipientFee:
        example: 100
        type: integer
      sender:
        example: 03abc...
        type: string
      status:
        example: payment_required
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
    type: object
  handlers.QuoteEntry:
    description: Quote for one recipient in batch
    properties:
      deliveryFee:
        example: 10
        type: integer
      messageBox:
        example: inbox
        type: string
      recipient:
        example: 03abc...
        type: string
      recipientFee:
        example: 100
        type: integer
      status:
        example: payment_required
        type: string
    type: object
  handlers.QuoteMultiResponse:
    description: Response containing quotes for multiple recipients
    properties:
      blockedRecipients:
        items:
          type: string
        type: array
      description:
        example: Message delivery quotes generated for 2 recipients.
        type: string
      quotesByRecipient:
        items:
          $ref: '#/definitions/handlers.QuoteEntry'
        type: array
      status:
        example: success
        type: string
      totals:
        $ref: '#/definitions/handlers.QuoteTotals'
    type: object
...
  handlers.QuoteTotals:
    description: Totals for batch quote
    properties:
      deliveryFees:
        example: 20
        type: integer
      recipientFees:
        example: 200
        type: integer
      totalForPayableRecipients:
        example: 220
        type: integer
    type: object
  handlers.RegisterDeviceRequest:
    description: Request to register a device for push notifications
    properties:
      deviceId:
        type: string
      fcmToken:
        type: string
      platform:
        type: string
    type: object
  handlers.RegisterDeviceResponse:
    description: Response after registering a device
    properties:
      deviceId:
        example: 1
        type: integer
      message:
        example: Device registered successfully for push notifications
        type: string
      status:
        example: success
        type: string
    type: object
  handlers.SendMessageRequest:
    type: object
  handlers.SendMessageResponse:
    description: Response after sending message(s)
    properties:
      message:
        example: Your message has been sent to 1 recipient(s).
        type: string
      results:
        items:
          $ref: '#/definitions/handlers.SendMessageResult'
        type: array
      status:
        example: success
        type: string
    type: object
  handlers.SendMessageResult:
    description: Result for a single recipient
    properties:
      messageId:
        example: msg-123
        type: string
      recipient:
        example: 03abc...
        type: string
    type: object
  handlers.SetPermissionRequest:
    description: Request to set a permission
    properties:
      messageBox:
        example: inbox
        type: string
      recipientFee:
        example: 100
        type: integer
      sender:
        example: 03abc...
        type: string
    type: object
  handlers.SetPermissionResponse:
    description: Response after setting a permission
    properties:
      description:
        example: Messages from sender to inbox now require 100 satoshis.
        type: string
      status:
        example: success
        type: string
    type: object
  handlers.SuccessResponse:
    description: Simple success response
    properties:
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for message delivery, retrieval, acknowledgement and permissions.
    Uses BRC-31/BRC-104 mutual authentication.
  title: MessageBox Server API
  version: 1.0.0
paths:
  /acknowledgeMessage:
    post:
      consumes:
      - application/json
      description: Removes acknowledged messages from the database for the authenticated
        identity. Used after a client has received and processed messages.
      parameters:
      - description: Message IDs to acknowledge
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AcknowledgeMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Acknowledge receipt of messages
      tags:
      - Messages
  /devices:
    get:
      description: Returns all devices registered for push notifications for the authenticated
        identity.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListDevicesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: List registered devices
      tags:
      - Devices
  /listMessages:
    post:
      consumes:
      - application/json
      description: Returns all stored messages for the specified messageBox belonging
        to the authenticated identity. If the box does not exist or has no messages,
        an empty array is returned.
      parameters:
      - description: Message box to list messages from
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ListMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListMessagesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Retrieve messages from a message box
      tags:
      - Messages
  /permissions/get:
    get:
      description: Retrieves the permission setting for a specific sender or box-wide
        default for the authenticated identity.
      parameters:
      - description: Name of the message box
        in: query
        name: messageBox
        required: true
        type: string
      - description: Sender's public key (omit for box-wide setting)
        in: query
        name: sender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetPermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Get a message permission
      tags:
      - Permissions
  /permissions/list:
    get:
      description: Returns all permission settings for the authenticated identity,
        optionally filtered by message box.
      parameters:
      - description: Filter by message box name
        in: query
        name: messageBox
        type: string
      - description: Maximum number of results (1-1000, default 100)
        in: query
        name: limit
        type: integer
      - description: Number of results to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: 'Sort order: ''asc'' or ''desc'' (default ''desc'')'
        in: query
        name: createdAtOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: List message permissions
      tags:
      - Permissions
  /permissions/quote:
    get:
      description: Returns fee information for sending a message to one or more recipients.
        Single recipient returns QuoteSingleResponse, multiple recipients returns
        QuoteMultiResponse.
      parameters:
      - description: Recipient public key (can be repeated for multiple recipients)
        in: query
        name: recipient
        required: true
        type: string
      - description: Name of the message box
        in: query
        name: messageBox
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Multiple recipient quote
          schema:
            $ref: '#/definitions/handlers.QuoteMultiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Get a delivery quote
      tags:
      - Permissions
  /permissions/set:
    post:
      consumes:
      - application/json
      description: Sets fee requirements for receiving messages. Use recipientFee=0
        for free, recipientFee=-1 to block, or a positive value for required payment
        in satoshis. Omit sender for box-wide defaults.
      parameters:
      - description: Permission settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SetPermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Set a message permission
      tags:
      - Permissions
  /registerDevice:
    post:
      consumes:
      - application/json
      description: Registers a device with an FCM token for receiving push notifications.
        Supports iOS, Android, and web platforms.
      parameters:
      - description: Device registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RegisterDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Register a device for push notifications
      tags:
      - Devices
  /sendMessage:
    post:
      consumes:
      - application/json
      description: Inserts a message into the target recipient's message box. Supports
        single or multiple recipients. Payment may be required depending on recipient's
        fee settings.
      parameters:
      - description: Message to send
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SendMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.DeliveryBlockedError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Send a message to recipient(s)
      tags:
      - Messages
securityDefinitions:
  BSVAuth:
    description: BRC-31/BRC-104 mutual authentication. Requires multiple x-bsv-auth-*
      headers (identity-key, nonce, signature, etc.)
    in: header
    name: x-bsv-auth-identity-key
    type: apiKey
swagger: "2.0"